fix(database): split out mongo file
